[refactor] move result to its own module
